refactor: Move Mono.Cecil extension methods not specific to either apireference or commandlinehelp to Common assembly
