--- src/MdDocs.ApiReference/Model/Constants.cs
@@ -1,17 +1,0 @@
-﻿namespace Grynwald.MdDocs.ApiReference.Model
-{
-    internal static class Constants
-    {
-        public const string ExtensionAttributeFullName = "System.Runtime.CompilerServices.ExtensionAttribute";
-        public const string StringFullName = "System.String";
-        public const string FlagsAttributeFullName = "System.FlagsAttribute";
-        public const string ObjectFullName = "System.Object";
-        public const string ValueTypeFullName = "System.ValueType";
-        public const string EnumFullName = "System.Enum";
-        public const string Int32FullName = "System.Int32";
-        public const string DefaultMemberAttributeFullName = "System.Reflection.DefaultMemberAttribute";
-        public const string IsReadOnlyAttributeFullName = "System.Runtime.CompilerServices.IsReadOnlyAttribute";
-        public const string ObsoleteAttributeFullName = "System.ObsoleteAttribute";
-        public const string ParamArrayAttributeFullName = "System.ParamArrayAttribute";
-    }
-}
