+++ src/MdDocs.ApiReference/Model/SimpleMemberDocumentation.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using Grynwald.MdDocs.ApiReference.Model.XmlDocs;
+using Grynwald.MdDocs.Common;
 using Grynwald.Utilities.Collections;
 using Mono.Cecil;
