+++ src/MdDocs.CommandLineHelp/CommandLineParserTypeNames.cs
@@ -1,17 +1,15 @@
 ﻿#pragma warning disable IDE1006 // Naming Styles: Ignore hint that const strings should be prefixed with 's_' as all these fields are public
 namespace Grynwald.MdDocs.CommandLineHelp
 {
-    internal static class Constants
+    /// <summary>
+    /// Defines constants for names of type names defined by the "CommandLineParser" library
+    /// </summary>
+    internal static class CommandLineParserTypeNames
     {
         public const string VerbAttributeFullName = "CommandLine.VerbAttribute";
         public const string OptionAttributeFullName = "CommandLine.OptionAttribute";
         public const string ValueAttributeFullName = "CommandLine.ValueAttribute";
         public const string AssemblyUsageAttributeFullName = "CommandLine.Text.AssemblyUsageAttribute";
-
-        public const string AssemblyTitleAttributeFullName = "System.Reflection.AssemblyTitleAttribute";
-        public const string AssemblyInformationalVersionAttribute = "System.Reflection.AssemblyInformationalVersionAttribute";
-
-        public const string BooleanFullName = "System.Boolean";
     }
 }
 #pragma warning restore IDE1006 // Naming Styles
