+++ src/MdDocs.CommandLineHelp/Grynwald.MdDocs.CommandLineHelp.csproj
@@ -23,4 +23,10 @@
     <ProjectReference Include="..\MdDocs.Common\Grynwald.MdDocs.Common.csproj" />
   </ItemGroup>
   
+
+
+  <ItemGroup>
+    <Compile Include="../shared/Nullable.cs" />
+  </ItemGroup>
+
 </Project>
