+++ src/MdDocs.CommandLineHelp/Model/_Application/MultiCommandApplicationDocumentation.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using Grynwald.MdDocs.Common;
 using Microsoft.Extensions.Logging;
 using Mono.Cecil;
 
@@ -42,8 +43,8 @@ namespace Grynwald.MdDocs.CommandLineHelp.Model
         {
             return definition.MainModule.Types
                 .Where(x => !x.IsAbstract)
-                .WithAttribute(Constants.VerbAttributeFullName)
-                .Where(x => !x.GetAttribute(Constants.VerbAttributeFullName).GetPropertyValueOrDefault<bool>("Hidden"))
+                .WithAttribute(CommandLineParserTypeNames.VerbAttributeFullName)
+                .Where(x => !x.GetAttribute(CommandLineParserTypeNames.VerbAttributeFullName).GetPropertyValueOrDefault<bool>("Hidden"))
                 .Select(type => CommandDocumentation.FromTypeDefinition(this, type, logger))
                 .ToArray();
         }
