+++ src/MdDocs.Common/_Extensions/_Cecil/CustomAttributeExtensions.cs
@@ -1,11 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Diagnostics.CodeAnalysis;
+﻿using System.Diagnostics.CodeAnalysis;
 using System.Linq;
-using System.Text;
 using Mono.Cecil;
 
-namespace Grynwald.MdDocs.CommandLineHelp
+namespace Grynwald.MdDocs.Common
 {
     public static class CustomAttributeExtensions
     {
