+++ src/MdDocs.Common/_Extensions/_Cecil/FieldDefinitionExtensions.cs
@@ -2,28 +2,24 @@
 using System.Linq;
 using Mono.Cecil;
 
-namespace Grynwald.MdDocs.ApiReference.Model
+namespace Grynwald.MdDocs.Common
 {
-    internal static class MethodDefinitionExtensions
+    public static class FieldDefinitionExtensions
     {
         /// <summary>
-        /// Gets a methods's public custom attributes excluding attributes emitted by the C# compiler not relevant for the user.
+        /// Gets a fields's public custom attributes excluding attributes emitted by the C# compiler not relevant for the user.
         /// </summary>
         /// <returns>
         /// Returns all attributes except:
-        /// <list type="bullet">
-        ///     <item><c>ExtensionAttribute</c> </item>
+        /// <list type="bullet">        
         ///     <item>non-public Attribute types</item>
         /// </list>
         /// </returns>
-        public static IEnumerable<CustomAttribute> GetCustomAttributes(this MethodDefinition parameter)
+        public static IEnumerable<CustomAttribute> GetCustomAttributes(this FieldDefinition parameter)
         {
             return parameter.CustomAttributes
                 .Where(attribute =>
                 {
-                    if (attribute.AttributeType.FullName == Constants.ExtensionAttributeFullName)
-                        return false;
-
                     if (!attribute.AttributeType.Resolve().IsPublic)
                         return false;
