+++ src/MdDocs.Common/_Extensions/_Cecil/MethodDefinitionExtensions.cs
@@ -2,24 +2,28 @@
 using System.Linq;
 using Mono.Cecil;
 
-namespace Grynwald.MdDocs.ApiReference.Model
+namespace Grynwald.MdDocs.Common
 {
-    public static class FieldDefinitionExtensions
+    public static class MethodDefinitionExtensions
     {
         /// <summary>
-        /// Gets a fields's public custom attributes excluding attributes emitted by the C# compiler not relevant for the user.
+        /// Gets a methods's public custom attributes excluding attributes emitted by the C# compiler not relevant for the user.
         /// </summary>
         /// <returns>
         /// Returns all attributes except:
-        /// <list type="bullet">        
+        /// <list type="bullet">
+        ///     <item><c>ExtensionAttribute</c> </item>
         ///     <item>non-public Attribute types</item>
         /// </list>
         /// </returns>
-        public static IEnumerable<CustomAttribute> GetCustomAttributes(this FieldDefinition parameter)
+        public static IEnumerable<CustomAttribute> GetCustomAttributes(this MethodDefinition parameter)
         {
             return parameter.CustomAttributes
                 .Where(attribute =>
                 {
+                    if (attribute.AttributeType.FullName == SystemTypeNames.ExtensionAttributeFullName)
+                        return false;
+
                     if (!attribute.AttributeType.Resolve().IsPublic)
                         return false;
